1.4 テストコードの記述
除算メソッドのテスト

+++ junit-practice-introduction/chapter01/junit-tutorial/src/junit/tutorial/Calculator.java
@@ -5,7 +5,7 @@ public class Calculator {
 		return x * y;
 	}
 	
-	public int devide(int x, int y) {
-		return x / y;
+	public float devide(int x, int y) {
+		return (float) x / (float) y;
 	}
 }
